Fixed postProcessHtml option
Current usage of postProcessHtml option incorrectly assigns renderedRoutes.html to the output of this._options.postProcessHtml(). However, renderedRoutes is an array of route objects with both route and html properties, so the assignment of renderedRoutes.html doesn't make sense here. This proposed change allows users to pass a function as the postProcessHtml option in the webpack plugin. This function is responsible for processing the array of route objects and returning it as a similar array, processedRoutes, to be used in the next .then() statement on line 47.